test(card): isolate ask-user lifecycle coverage
